catalog/tests/functional/requirements.txt
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pydantic==1.9.0
orjson==3.9.2
aiohttp==3.8.4
redis==4.4.2
elasticsearch[async]==7.9.1